pyproject.toml
[project]
name = "bd-automation"
version = "0.1.0"
description = "Business development automation suite for parking and food packaging import streams"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "thefuzz[speedup]>=0.22.1",
    "googlemaps>=4.10.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "modules"]

[tool.pytest.ini_options]
testpaths = ["modules"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
